# distsqlrun: add countRows processor