This is a simple processor that counts the number of rows it receives.
It's about 10x faster than using the generalized aggregator for this
purpose.

Before:
```
BenchmarkCountRows-8         500           4012735 ns/op         199.37 MB/s
```

After:
```
BenchmarkCountRows-8        3000            549780 ns/op        1455.13 MB/s
```

Release note: None