@RaduBerinde 2018-06-19
Wow, this speed-up is depressing. Any idea what's so inefficient in the general aggregator?

---

Review status: :shipit: complete! 0 of 0 LGTMs obtained

---



*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26828#-:-LFNskPT8A-DvB4qdmB4:b-83lbek)*
<!-- Sent from Reviewable.io -->