@jordanlewis 2018-06-19
It's a bunch of stuff. It's hard to tell via pprof, though. Seems to be death by ~a thousand~ several cuts. The normal implementation is pretty efficient and shortcuts when there are no columns, but there's still a bunch of work that gets done for no reason, lots of extra indexing into arrays, some copies, some spurious map lookups, that kind of thing.

I don't understand how this adds up to a 10x speedup though.