@jordanlewis 2018-06-19
Alfonso had the wise suggestion to make this slightly more general, to handle aggregates with no grouping, which is indeed a more general special case. I think I'll leave that for a separate PR, though - I have a feeling that doing so might erase these gains, and COUNT(*) is a classic special case that might well deserve a specialized implementation. Then again, this doesn't really affect overall query time too much that I can tell, since we're limited by our scan implementation.