@jordanlewis 2018-06-19
It seems like the slow stuff in the general aggregator is all of the indirection to aggregateFuncHolders. Maybe the pointer lookups there are what gets us. Perhaps we should be storing a slice of concrete aggregateFuncHolders rather than pointers to them. I can experiment with that also.