@solongordon 2018-06-19
😮
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>
Interested to see what comes out of your experiments. It feels like there must be a general lesson to be learned here.

---

Review status: :shipit: complete! 0 of 0 LGTMs obtained (and 1 stale)

---



*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26828#-:-LFOM0901cBw5zEEuAdX:b14kag5)*
<!-- Sent from Reviewable.io -->