@jordanlewis 2018-10-02
Resurrected from the dead.

```
name                old time/op    new time/op    delta
Count/Cockroach-24    33.3ms ± 4%    29.5ms ± 1%  -11.56%  (p=0.000 n=9+10)

name                old alloc/op   new alloc/op   delta
Count/Cockroach-24    3.77MB ± 0%    3.76MB ± 0%   -0.22%  (p=0.000 n=10+9)

name                old allocs/op  new allocs/op  delta
Count/Cockroach-24     1.00k ± 1%     0.97k ± 1%   -2.47%  (p=0.000 n=9+8)
```